# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentio SMTP API",
    "description": "Multi-tenant SMTP platform API",
    "contact": {
      "name": "Truespar Team"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/reports/dmarc": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "list_dmarc_reports",
        "parameters": [
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Vec_DmarcReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reports/dmarc/{id}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "get_dmarc_report",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_DmarcReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reports/fbl": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "list_fbl_reports",
        "parameters": [
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Vec_FblReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reports/fbl/{id}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "get_fbl_report",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_FblReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reports/fbl/{id}/process": {
      "post": {
        "tags": [
          "Reports"
        ],
        "operationId": "process_fbl_report",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_FblReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reports/tlsrpt": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "list_tlsrpt_reports",
        "parameters": [
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Vec_TlsrptReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reports/tlsrpt/{id}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "get_tlsrpt_report",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TlsrptReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/spam/train/ham": {
      "post": {
        "tags": [
          "Spam Training"
        ],
        "operationId": "train_ham",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message learned as ham"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/spam/train/spam": {
      "post": {
        "tags": [
          "Spam Training"
        ],
        "operationId": "train_spam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message learned as spam"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplaintType": {
        "type": "string",
        "enum": [
          "abuse",
          "fraud",
          "virus",
          "other",
          "not-spam"
        ]
      },
...
      "DataResponse_DmarcReportResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "domain_id",
              "direction",
              "report_id",
              "date_begin",
              "date_end",
              "total_count",
              "dkim_pass",
              "dkim_fail",
              "spf_pass",
              "spf_fail",
              "dmarc_pass",
              "dmarc_fail",
              "created_at"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "date_begin": {
                "type": "string",
                "format": "date-time"
              },
              "date_end": {
                "type": "string",
                "format": "date-time"
              },
              "direction": {
                "$ref": "#/components/schemas/MessageDirection"
              },
              "dkim_fail": {
                "type": "integer",
                "format": "int32"
              },
              "dkim_pass": {
                "type": "integer",
                "format": "int32"
              },
              "dmarc_fail": {
                "type": "integer",
                "format": "int32"
              },
              "dmarc_pass": {
                "type": "integer",
                "format": "int32"
              },
              "domain_id": {
                "$ref": "#/components/schemas/DomainId"
              },
              "id": {
                "$ref": "#/components/schemas/DmarcReportId"
              },
              "org_name": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "report_id": {
                "type": "string"
              },
              "source_ip": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "spf_fail": {
                "type": "integer",
                "format": "int32"
              },
              "spf_pass": {
                "type": "integer",
                "format": "int32"
              },
              "total_count": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        }
      },
...
      "DataResponse_FblReportResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "complained_recipient",
              "complaint_type",
              "auto_suppressed",
              "created_at"
            ],
            "properties": {
              "arrival_date": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "auto_suppressed": {
                "type": "boolean"
              },
              "complained_recipient": {
                "type": "string"
              },
              "complaint_type": {
                "$ref": "#/components/schemas/ComplaintType"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "feedback_type": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "id": {
                "$ref": "#/components/schemas/FblReportId"
              },
              "processed_at": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "source_ip": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        }
      },
...
      "DataResponse_Vec_DmarcReportResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "domain_id",
                "direction",
                "report_id",
                "date_begin",
                "date_end",
                "total_count",
                "dkim_pass",
                "dkim_fail",
                "spf_pass",
                "spf_fail",
                "dmarc_pass",
                "dmarc_fail",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "date_begin": {
                  "type": "string",
                  "format": "date-time"
                },
                "date_end": {
                  "type": "string",
                  "format": "date-time"
                },
                "direction": {
                  "$ref": "#/components/schemas/MessageDirection"
                },
                "dkim_fail": {
                  "type": "integer",
                  "format": "int32"
                },
                "dkim_pass": {
                  "type": "integer",
                  "format": "int32"
                },
                "dmarc_fail": {
                  "type": "integer",
                  "format": "int32"
                },
                "dmarc_pass": {
                  "type": "integer",
                  "format": "int32"
                },
                "domain_id": {
                  "$ref": "#/components/schemas/DomainId"
                },
                "id": {
                  "$ref": "#/components/schemas/DmarcReportId"
                },
                "org_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "report_id": {
                  "type": "string"
                },
                "source_ip": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "spf_fail": {
                  "type": "integer",
                  "format": "int32"
                },
                "spf_pass": {
                  "type": "integer",
                  "format": "int32"
                },
                "total_count": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
...
      "DataResponse_Vec_FblReportResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "complained_recipient",
                "complaint_type",
                "auto_suppressed",
                "created_at"
              ],
              "properties": {
                "arrival_date": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "auto_suppressed": {
                  "type": "boolean"
                },
                "complained_recipient": {
                  "type": "string"
                },
                "complaint_type": {
                  "$ref": "#/components/schemas/ComplaintType"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "feedback_type": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "$ref": "#/components/schemas/FblReportId"
                },
                "processed_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "source_ip": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          }
        }
      },
...
      "DmarcReportId": {
        "type": "string",
        "format": "uuid"
      },
...
      "DomainId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "FblReportId": {
        "type": "string",
        "format": "uuid"
      },
...
      "MessageDirection": {
        "type": "string",
        "enum": [
          "inbound",
          "outbound"
        ]
      },
...
      "TrainRequest": {
        "type": "object",
        "required": [
          "raw_message"
        ],
        "properties": {
          "raw_message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
...
}